Refactor Bknd configuration and imports
This commit reorganizes imports in `bknd.ts` for better readability and
introduces a significant change to the application's mode. The `mode` is
now dynamically set to 'db' in development environments and 'code'
otherwise, enabling easier local development with database-driven
configurations. Additionally, a `manager.secrets` option is added to the
configuration, allowing environment variables to be accessed. The
`index.tsx` file also sees minor import reordering.